+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/generics/GenericFunctionTypingError.scala
@@ -1,0 +1,15 @@
+package pl.touk.nussknacker.engine.api.generics
+
+import cats.data.NonEmptyList
+
+sealed trait GenericFunctionTypingError
+
+object GenericFunctionTypingError {
+  case object ArgumentTypeError extends GenericFunctionTypingError
+
+  trait CustomError extends GenericFunctionTypingError {
+    def message: String
+  }
+
+  case class OtherError(message: String) extends CustomError
+}