+++ components-api/src/test/scala/pl/touk/nussknacker/engine/api/typed/TypingResultSpec.scala
@@ -6,13 +6,12 @@ import org.scalatest.matchers.should.Matchers
 import pl.touk.nussknacker.engine.api.typed.supertype.{CommonSupertypeFinder, NumberTypesPromotionStrategy, SupertypeClassResolutionStrategy}
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedClass, TypedNull, TypedObjectTypingResult, TypedUnion, TypingResult, Unknown}
 
+import java.util
 import scala.collection.immutable.ListMap
 
 class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with Inside {
 
   private val commonSuperTypeFinder = new CommonSupertypeFinder(SupertypeClassResolutionStrategy.Intersection, true)
-
-  implicit val numberTypesPromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ForMathOperation
 
   private def typeMap(args: (String, TypingResult)*) = TypedObjectTypingResult(args.toList)
 
@@ -25,7 +24,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
     ) shouldBe true
 
     typeMap("field1" -> Typed[String]).canBeSubclassOf(
-          typeMap("field1" -> Typed[String], "field2" -> Typed[Int])
+      typeMap("field1" -> Typed[String], "field2" -> Typed[Int])
     ) shouldBe false
 
     typeMap("field1" -> Typed[Int]).canBeSubclassOf(
@@ -86,6 +85,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("find common supertype for simple types enabled strictTypeChecking") {
+    implicit val numberTypesPromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ForMathOperation
     commonSuperTypeFinder.commonSupertype(Typed[String], Typed[String]) shouldEqual Typed[String]
     commonSuperTypeFinder.commonSupertype(Typed[java.lang.Integer], Typed[java.lang.Double]) shouldEqual Typed[java.lang.Double]
     commonSuperTypeFinder.commonSupertype(Typed[Int], Typed[Double]) shouldEqual Typed[java.lang.Double]
@@ -97,6 +97,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("find common supertype for simple types disabled strictTypeChecking") {
+    implicit val numberTypesPromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ForMathOperation
     val typeFinderDisabledStrictTypeChecking = new CommonSupertypeFinder(SupertypeClassResolutionStrategy.Intersection, false)
     typeFinderDisabledStrictTypeChecking.commonSupertype(Typed[String], Typed[String]) shouldEqual Typed[String]
     typeFinderDisabledStrictTypeChecking.commonSupertype(Typed[java.lang.Integer], Typed[java.lang.Double]) shouldEqual Typed[java.lang.Double]
@@ -109,6 +110,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("find special types") {
+    implicit val numberTypesPromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ForMathOperation
     commonSuperTypeFinder.commonSupertype(Unknown, Unknown) shouldEqual Unknown
     commonSuperTypeFinder.commonSupertype(Unknown, Typed[Long]) shouldEqual Unknown
 
@@ -123,6 +125,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("find common supertype for complex types with inheritance in classes hierarchy") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
     import ClassHierarchy._
     commonSuperTypeFinder.commonSupertype(Typed[Dog], Typed[Pet]) shouldEqual Typed[Pet]
     commonSuperTypeFinder.commonSupertype(Typed[Dog], Typed[Cat]) shouldEqual Typed[Pet]
@@ -132,6 +135,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("find common supertype for complex types with inheritance in interfaces hierarchy") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
     import InterfaceHierarchy._
     commonSuperTypeFinder.commonSupertype(Typed[Dog], Typed[Pet]) shouldEqual Typed[Pet]
     commonSuperTypeFinder.commonSupertype(Typed[Dog], Typed[Cat]) shouldEqual Typed[Pet]
@@ -141,6 +145,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("find common supertype for complex types with inheritance in mixins hierarchy") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
     import HierarchyInMixins._
     commonSuperTypeFinder.commonSupertype(Typed[Dog], Typed[Pet]) shouldEqual Typed[Pet]
     commonSuperTypeFinder.commonSupertype(Typed[Dog], Typed[Cat]) shouldEqual Typed[Pet]
@@ -149,7 +154,19 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
     commonSuperTypeFinder.commonSupertype(Typed(Typed[Dog], Typed[Cactus]), Typed[Cat]) shouldEqual Typed[Pet]
   }
 
+  test("common supertype with generics") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
+    commonSuperTypeFinder.commonSupertype(Typed.fromDetailedType[Comparable[Number]], Typed.fromDetailedType[Integer]) shouldEqual Typed.fromDetailedType[Comparable[Number]]
+    commonSuperTypeFinder.commonSupertype(Typed.fromDetailedType[Integer], Typed.fromDetailedType[Comparable[Number]]) shouldEqual Typed.fromDetailedType[Comparable[Number]]
+    commonSuperTypeFinder.commonSupertype(Typed.fromDetailedType[util.List[Integer]], Typed.fromDetailedType[util.List[Number]]) shouldEqual Typed.fromDetailedType[util.List[Number]]
+    commonSuperTypeFinder.commonSupertype(Typed.fromDetailedType[util.List[Integer]], Typed.fromDetailedType[util.Collection[Number]]) shouldEqual Typed.fromDetailedType[util.Collection[Number]]
+    // below weird examples which will work
+    commonSuperTypeFinder.commonSupertype(Typed.fromDetailedType[util.List[Number]], Typed.fromDetailedType[util.Collection[Integer]]) shouldEqual Typed.fromDetailedType[util.Collection[Number]]
+    commonSuperTypeFinder.commonSupertype(Typed.fromDetailedType[util.List[String]], Typed.fromDetailedType[util.Collection[Integer]]) shouldEqual Typed.genericTypeClass[util.Collection[_]](List(Unknown))
+  }
+
   test("common supertype with union of not matching classes strategy with enabled strictTypeChecking") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
     import ClassHierarchy._
     val unionFinder = new CommonSupertypeFinder(SupertypeClassResolutionStrategy.Union, true)
     unionFinder.commonSupertype(Typed[Dog], Typed[Cactus]) shouldEqual Typed(Typed[Dog], Typed[Cactus])
@@ -158,6 +175,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("common supertype with union of not matching classes strategy with disabled strictTypeChecking") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
     import ClassHierarchy._
     val unionFinder = new CommonSupertypeFinder(SupertypeClassResolutionStrategy.Union, false)
     unionFinder.commonSupertype(Typed[Dog], Typed[Cactus]) shouldEqual Typed(Typed[Dog], Typed[Cactus])
@@ -207,6 +225,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("determinate if can be superclass for objects with value") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
     val unionFinder = new CommonSupertypeFinder(SupertypeClassResolutionStrategy.Union, false)
     unionFinder.commonSupertype(Typed.fromInstance(65), Typed.fromInstance(65)) shouldBe Typed.fromInstance(65)
     unionFinder.commonSupertype(Typed.fromInstance(91), Typed.fromInstance(35)) shouldBe Typed.typedClass[Int]
@@ -214,6 +233,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("should calculate supertype for objects with value when strict type checking is on") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
     val unionFinder = new CommonSupertypeFinder(SupertypeClassResolutionStrategy.Union, true)
     unionFinder.commonSupertype(Typed.fromInstance(65), Typed.fromInstance(65)) shouldBe Typed.fromInstance(65)
     unionFinder.commonSupertype(Typed.fromInstance(91), Typed.fromInstance(35)) shouldBe Typed.typedClass[Int]
@@ -221,6 +241,7 @@ class TypingResultSpec extends AnyFunSuite with Matchers with OptionValues with 
   }
 
   test("should calculate supertype for null") {
+    implicit val toSupertypePromotionStrategy: NumberTypesPromotionStrategy = NumberTypesPromotionStrategy.ToSupertype
     val unionFinder = new CommonSupertypeFinder(SupertypeClassResolutionStrategy.Union, false)
     unionFinder.commonSupertype(TypedNull, TypedNull) shouldBe TypedNull
     unionFinder.commonSupertype(TypedNull, Typed[String]) shouldBe Typed[String]