+++ engine/flink/management/dev-model/src/main/scala/pl/touk/nussknacker/engine/management/sample/global/GenericHelperFunction.scala
@@ -1,11 +1,13 @@
 package pl.touk.nussknacker.engine.management.sample.global
 
-import cats.data.{NonEmptyList, Validated, ValidatedNel}
+import cats.data.{NonEmptyList, ValidatedNel}
 import cats.implicits.catsSyntaxValidatedId
-import pl.touk.nussknacker.engine.api.generics.{ArgumentTypeError, ExpressionParseError, GenericFunctionError, GenericType, Signature, TypingFunction}
+import pl.touk.nussknacker.engine.api.generics.{GenericFunctionTypingError, GenericType, MethodTypeInfo, Parameter, Signature, TypingFunction}
 import pl.touk.nussknacker.engine.api.Documentation
-import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedClass, TypedObjectTypingResult, TypedObjectWithValue, TypingResult, Unknown}
+import pl.touk.nussknacker.engine.api.generics.GenericFunctionTypingError.ArgumentTypeError
+import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedClass, TypedObjectTypingResult, TypedObjectWithValue, TypingResult}
 
+import scala.annotation.varargs
 import scala.jdk.CollectionConverters.collectionAsScalaIterableConverter
 
 object GenericHelperFunction {
@@ -17,17 +19,10 @@ object GenericHelperFunction {
   private class HeadHelper extends TypingFunction {
     private val listClass = classOf[java.util.List[_]]
 
-    private def error(arguments: List[TypingResult]) = {
-      new ArgumentTypeError(
-        new Signature("head", arguments, None),
-        List(new Signature("head", List(Typed.fromDetailedType[java.util.List[Object]]), None))
-      )
-    }
-
-    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[ExpressionParseError, TypingResult] = arguments match {
+    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[GenericFunctionTypingError, TypingResult] = arguments match {
       case TypedClass(`listClass`, t :: Nil) :: Nil => t.validNel
       case TypedClass(`listClass`, _) :: Nil => throw new AssertionError("Lists must have one parameter")
-      case _ => error(arguments).invalidNel
+      case _ => GenericFunctionTypingError.ArgumentTypeError.invalidNel
     }
   }
 
@@ -41,34 +36,24 @@ object GenericHelperFunction {
     private val expectedArgument = Typed(Typed.fromInstance("String"), Typed.fromInstance("Double"), Typed.fromInstance("Int"))
     private val expectedResult = Typed(Typed[String], Typed[Double], Typed[Int])
 
-    private def error(arguments: List[TypingResult]) = {
-      new ArgumentTypeError(
-        new Signature("exampleOfType", arguments, None),
-        List(new Signature("exampleOfType", List(expectedArgument), None))
-      )
-    }
+    override def signatures: Option[NonEmptyList[MethodTypeInfo]] =
+      Some(NonEmptyList.one(MethodTypeInfo(List(Parameter("typeName", expectedArgument)), None, expectedResult)))
 
-    override def staticParameters(): Option[List[(String, TypingResult)]] =
-      Some(List(("typeName", expectedArgument)))
-
-    override def staticResult(): Option[TypingResult] =
-      Some(expectedResult)
-
-    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[ExpressionParseError, TypingResult] = arguments match {
+    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[GenericFunctionTypingError, TypingResult] = arguments match {
       case TypedObjectWithValue(TypedClass(`stringClass`, Nil), typ: String) :: Nil => typ match {
         case "String" => Typed[String].validNel
         case "Int" => Typed[Int].validNel
         case "Double" => Typed[Double].validNel
-        case _ => new GenericFunctionError("Expected string with value 'String', 'Int' or 'Double'").invalidNel
+        case _ => GenericFunctionTypingError.ArgumentTypeError.invalidNel
       }
       case a@TypedObjectWithValue(TypedClass(`stringClass`, _), _) :: Nil =>
         throw new AssertionError(s"Found illegal type $a")
       case TypedClass(`stringClass`, Nil) :: Nil =>
-        new GenericFunctionError("Expected string with known value").invalidNel
+        GenericFunctionTypingError.OtherError("Expected string with known value").invalidNel
       case a@TypedClass(`stringClass`, _) :: Nil =>
         throw new AssertionError(s"Found illegal type $a")
       case _ =>
-        error(arguments).invalidNel
+        GenericFunctionTypingError.ArgumentTypeError.invalidNel
     }
   }
 
@@ -78,15 +63,13 @@ object GenericHelperFunction {
   def getFieldA(obj: Any): Any = ???
 
   private class RequiresFiledAHelper extends TypingFunction {
-    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[ExpressionParseError, TypingResult] = arguments match {
+    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[GenericFunctionTypingError, TypingResult] = arguments match {
       case TypedObjectTypingResult(fields, _, _) :: Nil => fields.get("a") match {
         case Some(x) => x.validNel
-        case None => new GenericFunctionError("Given object does not have field 'a'").invalidNel
+        case None => GenericFunctionTypingError.OtherError("Given object does not have field 'a'").invalidNel
       }
-      case _ :: Nil =>
-        new GenericFunctionError("Expected typed object").invalidNel
       case _ =>
-        new GenericFunctionError("Expected one argument").invalidNel
+        GenericFunctionTypingError.ArgumentTypeError.invalidNel
     }
   }
 
@@ -96,15 +79,37 @@ object GenericHelperFunction {
   def addFieldA(obj: Any): Any = ???
 
   private class AddFieldAHelper extends TypingFunction {
-    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[ExpressionParseError, TypingResult] = arguments match {
+    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[GenericFunctionTypingError, TypingResult] = arguments match {
       case TypedObjectTypingResult(fields, obj, info) :: Nil => fields.get("a") match {
-        case Some(_) => new GenericFunctionError("Given object already has field 'a'").invalidNel
+        case Some(_) => GenericFunctionTypingError.OtherError("Given object already has field 'a'").invalidNel
         case None => TypedObjectTypingResult(fields + ("a" -> Typed[Int]), obj, info).validNel
       }
-      case _ :: Nil =>
-        new GenericFunctionError("Expected typed object").invalidNel
       case _ =>
-        new GenericFunctionError("Expected one argument").invalidNel
+        GenericFunctionTypingError.ArgumentTypeError.invalidNel
+    }
+  }
+
+  @Documentation(description = "zips up to 3 numbers")
+  @GenericType(typingFunction = classOf[ZipHelper])
+  @varargs
+  def zip(x: Number*): Any = ???
+
+  private class ZipHelper extends TypingFunction {
+    override def signatures: Option[NonEmptyList[MethodTypeInfo]] =
+      Some(NonEmptyList.of(
+        MethodTypeInfo(Parameter("arg", Typed[Number]) :: Nil, None, Typed[Number]),
+        MethodTypeInfo(Parameter("left", Typed[Number]) :: Parameter("right", Typed[Number]) :: Nil, None, Typed[Number]),
+        MethodTypeInfo(Parameter("left", Typed[Number]) :: Parameter("mid", Typed[Number]) :: Parameter("right", Typed[Number]) :: Nil, None, Typed[Number])
+      ))
+
+    override def computeResultType(arguments: List[TypingResult]): ValidatedNel[GenericFunctionTypingError, TypingResult] = {
+      if (arguments.exists(!_.canBeSubclassOf(Typed[Number]))) return ArgumentTypeError.invalidNel
+      arguments match {
+        case t :: Nil => t.validNel
+        case l :: r :: Nil => TypedObjectTypingResult(List("left" -> l, "right" -> r)).validNel
+        case l :: m :: r :: Nil => TypedObjectTypingResult(List("left" -> l, "mid" -> m, "right" -> r)).validNel
+        case _ => ArgumentTypeError.invalidNel
+      }
     }
   }
 }