+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/types/EspTypeUtils.scala
@@ -2,15 +2,19 @@ package pl.touk.nussknacker.engine.types
 
 import java.lang.reflect._
 import java.util.Optional
-import cats.data.StateT
+import cats.data.{NonEmptyList, StateT}
+import cats.data.Validated.Invalid
 import cats.effect.IO
+import cats.implicits.catsSyntaxSemigroup
 import org.apache.commons.lang3.{ClassUtils, StringUtils}
-import pl.touk.nussknacker.engine.api.generics.{GenericType, TypingFunction}
+import pl.touk.nussknacker.engine.api.generics.{GenericType, MethodTypeInfo, Parameter, TypingFunction}
 import pl.touk.nussknacker.engine.api.process.PropertyFromGetterExtractionStrategy.{AddPropertyNextToGetter, DoNothing, ReplaceGetterWithProperty}
 import pl.touk.nussknacker.engine.api.process.{ClassExtractionSettings, VisibleMembersPredicate}
-import pl.touk.nussknacker.engine.api.typed.typing.{SingleTypingResult, Typed, TypedUnion, TypingResult, Unknown}
+import pl.touk.nussknacker.engine.api.typed.typing.{SingleTypingResult, Typed, TypedNull, TypedUnion, TypingResult, Unknown}
 import pl.touk.nussknacker.engine.api.{Documentation, ParamName}
-import pl.touk.nussknacker.engine.definition.TypeInfos.{ClazzDefinition, FunctionalMethodInfo, MethodInfo, Parameter, StaticMethodInfo}
+import pl.touk.nussknacker.engine.definition.TypeInfos.{ClazzDefinition, FunctionalMethodInfo, MethodInfo, StaticMethodInfo}
+
+import java.lang.annotation.Annotation
 
 object EspTypeUtils {
 
@@ -48,27 +52,42 @@ object EspTypeUtils {
     }
     val publicMethods = clazz.getMethods.toList ++ additionalMethods
 
-    val filteredMethods =
+    val methods =
       if (staticMethodsAndFields) publicMethods.filter(membersPredicate.shouldBeVisible).filter(m => Modifier.isStatic(m.getModifiers))
       else publicMethods.filter(membersPredicate.shouldBeVisible).filter(m => !Modifier.isStatic(m.getModifiers))
+
+    // "varargs" annotation generates two methods - one with scala style varArgs
+    // and one with java style varargs. We want only the second one so we have
+    // to filter them.
+    val filteredMethods = methods.filter(extractJavaVersionOfVarArgMethod(_).isEmpty)
 
     val methodNameAndInfoList = filteredMethods
       .flatMap(extractMethod(_))
 
-    deduplicateMethodsWithGenericReturnType(methodNameAndInfoList)
+    val staticMethodInfos = methodNameAndInfoList.filter(_._2.isInstanceOf[StaticMethodInfo]).asInstanceOf[List[(String, StaticMethodInfo)]]
+    val functionalMethodInfos = methodNameAndInfoList.filter(_._2.isInstanceOf[FunctionalMethodInfo])
+    val groupedFunctionalMethodInfos = functionalMethodInfos.groupBy(_._1).mapValues(_.map(_._2))
+
+    deduplicateMethodsWithGenericReturnType(staticMethodInfos)
+      .asInstanceOf[Map[String, List[MethodInfo]]]
+      .combine(groupedFunctionalMethodInfos)
   }
 
   //We have to filter here, not in ClassExtractionSettings, as we do e.g. boxed/unboxed mapping on TypedClass level...
   private def filterHiddenParameterAndReturnType(infos: Map[String, List[MethodInfo]])
                                                 (implicit settings: ClassExtractionSettings): Map[String, List[MethodInfo]] = {
-    def typeResultVisible(str: SingleTypingResult) = !settings.isHidden(str.objType.klass)
+    def typeResultVisible(t: TypingResult): Boolean = t match {
+      case str: SingleTypingResult =>
+        !settings.isHidden(str.objType.klass) && str.objType.params.forall(typeResultVisible)
+      case TypedUnion(ts) => ts.forall(typeResultVisible)
+      case TypedNull => true
+      case Unknown => true
+    }
     def filterOneMethod(methodInfo: MethodInfo): Boolean = {
-      (methodInfo.staticParameters.map(_.refClazz) :+ methodInfo.staticResult).forall {
-        //TODO: handle arrays properly in ClassExtractionSettings
-        case e: SingleTypingResult => (methodInfo.varArgs && e.objType.klass.isArray) || typeResultVisible(e)
-        case TypedUnion(results) => results.forall(typeResultVisible)
-        case _ => true
-      }
+      val noVarArgTypes = methodInfo.signatures.toList.flatMap(_.noVarArgs).map(_.refClazz)
+      val varArgTypes = methodInfo.signatures.toList.flatMap(_.varArg.toList).map(_.refClazz)
+      val resultTypes = methodInfo.signatures.toList.map(_.result)
+      (noVarArgTypes ::: varArgTypes ::: resultTypes).forall(typeResultVisible)
     }
     infos.mapValuesNow(methodList => methodList.filter(filterOneMethod)).filter(_._2.nonEmpty)
   }
@@ -83,8 +102,8 @@ object EspTypeUtils {
       LocalDate toLocalDate()
     In our case the second one is correct
    */
-  private def deduplicateMethodsWithGenericReturnType(methodNameAndInfoList: List[(String, MethodInfo)]) = {
-    val groupedByNameAndParameters = methodNameAndInfoList.groupBy(mi => (mi._1, mi._2.staticParameters))
+  private def deduplicateMethodsWithGenericReturnType(methodNameAndInfoList: List[(String, StaticMethodInfo)]) = {
+    val groupedByNameAndParameters = methodNameAndInfoList.groupBy(mi => (mi._1, mi._2.signature.noVarArgs, mi._2.signature.varArg))
     groupedByNameAndParameters.toList.map {
       case (_, methodsForParams) =>
         /*
@@ -94,8 +113,8 @@ object EspTypeUtils {
          */
 
         methodsForParams.find { case (_, methodInfo) =>
-          methodsForParams.forall(mi => methodInfo.staticResult.canBeSubclassOf(mi._2.staticResult))
-        }.getOrElse(methodsForParams.minBy(_._2.staticResult.display))
+          methodsForParams.forall(mi => methodInfo.signature.result.canBeSubclassOf(mi._2.signature.result))
+        }.getOrElse(methodsForParams.minBy(_._2.signature.result.display))
     }.toGroupedMap
       //we sort only to avoid randomness
       .mapValuesNow(_.sortBy(_.toString))
@@ -119,9 +138,9 @@ object EspTypeUtils {
 
   private def extractMethod(method: Method)
                            (implicit settings: ClassExtractionSettings): List[(String, MethodInfo)] =
-    method.getAnnotation(classOf[GenericType]) match {
-      case null => extractRegularMethod(method)
-      case annotation => extractGenericMethod(method, annotation)
+    extractAnnotation(method, classOf[GenericType]) match {
+      case None => extractRegularMethod(method)
+      case Some(annotation) => extractGenericMethod(method, annotation)
     }
 
   private def getTypeFunctionInstanceFromAnnotation(method: Method, genericType: GenericType): TypingFunction = {
@@ -144,30 +163,23 @@ object EspTypeUtils {
   private def extractGenericMethod(method: Method, genericType: GenericType)
                                   (implicit settings: ClassExtractionSettings): List[(String, MethodInfo)] = {
     val typeFunctionInstance = getTypeFunctionInstanceFromAnnotation(method, genericType)
-    val parameterInfo = typeFunctionInstance.staticParameters()
-      .map(_.map{ case (name, typ) => Parameter(name, typ) })
-      .getOrElse(extractParameters(method))
-    val resultInfo = typeFunctionInstance.staticResult()
-      .getOrElse(extractMethodReturnType(method))
 
-    collectMethodNames(method).map(methodName => methodName -> FunctionalMethodInfo.fromParameterList(
+    val methodTypeInfo = extractGenericParameters(typeFunctionInstance, method)
+
+    collectMethodNames(method).map(methodName => methodName -> FunctionalMethodInfo(
       x => typeFunctionInstance.computeResultType(x),
-      parameterInfo,
-      resultInfo,
+      methodTypeInfo,
       methodName,
-      extractNussknackerDocs(method),
-      method.isVarArgs
+      extractNussknackerDocs(method)
     ))
   }
 
   private def extractRegularMethod(method: Method)
                                   (implicit settings: ClassExtractionSettings): List[(String, StaticMethodInfo)] =
-    collectMethodNames(method).map(methodName => methodName -> StaticMethodInfo.fromParameterList(
-      extractParameters(method),
-      extractMethodReturnType(method),
+    collectMethodNames(method).map(methodName => methodName -> StaticMethodInfo(
+      extractMethodTypeInfo(method),
       methodName,
-      extractNussknackerDocs(method),
-      method.isVarArgs
+      extractNussknackerDocs(method)
     ))
 
   private def extractPublicFields(clazz: Class[_], membersPredicate: VisibleMembersPredicate, staticMethodsAndFields: Boolean)
@@ -177,27 +189,40 @@ object EspTypeUtils {
       if(staticMethodsAndFields) interestingFields.filter(m => Modifier.isStatic(m.getModifiers))
       else interestingFields.filter(m => !Modifier.isStatic(m.getModifiers))
     fields.map { field =>
-      field.getName -> StaticMethodInfo.fromParameterList(
-        List.empty,
-        extractFieldReturnType(field),
+      field.getName -> StaticMethodInfo(
+        MethodTypeInfo(Nil, None, extractFieldReturnType(field)),
         field.getName,
-        extractNussknackerDocs(field),
-        varArgs = false
+        extractNussknackerDocs(field)
       )
     }.toMap
   }
 
   private def extractNussknackerDocs(accessibleObject: AccessibleObject): Option[String] = {
-    Option(accessibleObject.getAnnotation(classOf[Documentation])).map(_.description())
+    extractAnnotation(accessibleObject, classOf[Documentation]).map(_.description())
   }
 
-  private def extractParameters(method: Method): List[Parameter] = {
-    for {
+  private def extractGenericParameters(typingFunction: TypingFunction, method: Method): NonEmptyList[MethodTypeInfo] = {
+    val autoExtractedParameters = extractMethodTypeInfo(method)
+    val definedParametersOption = typingFunction.signatures.toList.flatMap(_.toList)
+
+    definedParametersOption
+      .map(MethodTypeInfoSubclassChecker.check(_, autoExtractedParameters))
+      .collect{ case Invalid(e) => e }
+      .foreach { x =>
+        val errorString = x.map(_.message).toList.mkString("; ")
+        throw new IllegalArgumentException(s"Generic function ${method.getName} has declared parameters that are incompatible with methods signature: $errorString")
+      }
+
+    NonEmptyList.fromList(definedParametersOption).getOrElse(NonEmptyList.one(autoExtractedParameters))
+  }
+
+  private def extractMethodTypeInfo(method: Method): MethodTypeInfo = {
+    MethodTypeInfo.fromList(for {
       param <- method.getParameters.toList
-      annotationOption = Option(param.getAnnotation(classOf[ParamName]))
+      annotationOption = extractAnnotation(param, classOf[ParamName])
       name = annotationOption.map(_.value).getOrElse(param.getName)
       paramType = extractParameterType(param)
-    } yield Parameter(name, paramType)
+    } yield Parameter(name, paramType), method.isVarArgs, extractMethodReturnType(method))
   }
 
   def extractParameterType(javaParam: java.lang.reflect.Parameter): TypingResult = {
@@ -256,6 +281,39 @@ object EspTypeUtils {
   private def extractGenericParams(paramsType: ParameterizedType, paramsRawType: Class[_]): TypingResult = {
     Typed.genericTypeClass(paramsRawType, paramsType.getActualTypeArguments.toList.map(p => extractClass(p).getOrElse(Unknown)))
   }
+
+  private def extractScalaVersionOfVarArgMethod(method: Method): Option[Method] = {
+    val obj = method.getDeclaringClass
+    val name = method.getName
+    val args = method.getParameterTypes.toList
+    args match {
+      case noVarArgs :+ varArg if method.isVarArgs && varArg.isArray =>
+        try {
+          Some(obj.getMethod(name, noVarArgs :+ classOf[Seq[_]]: _*))
+        } catch {
+          case _: NoSuchMethodException => None
+        }
+      case _ => None
+    }
+  }
+
+  private def extractJavaVersionOfVarArgMethod(method: Method): Option[Method] = {
+    method.getDeclaringClass.getMethods.find(m => m.isVarArgs && (m.getParameterTypes.toList match {
+      case noVarArgs :+ varArgArr if varArgArr.isArray =>
+        method.getParameterTypes.toList == noVarArgs :+ classOf[Seq[_]]
+      case _ => false
+    }))
+  }
+
+  // "varargs" annotation creates new function that has java style varArgs
+  // but it disregards annotations, so we have to look for original function
+  // to extract them.
+  private def extractAnnotation[T <: Annotation](obj: AnnotatedElement, annotationType: Class[T]): Option[T] =
+    Option(obj.getAnnotation(annotationType)).orElse(obj match {
+      case method: Method => extractScalaVersionOfVarArgMethod(method).flatMap(extractAnnotation(_, annotationType))
+      // TODO: Add new case for parameters.
+      case _ => None
+    })
 
   def companionObject[T](klazz: Class[T]): T = {
     klazz.getField("MODULE$").get(null).asInstanceOf[T]