+++ interpreter/src/test/java/pl/touk/nussknacker/engine/types/JavaClassWithFilteredMethod.java
@@ -1,0 +1,10 @@
+package pl.touk.nussknacker.engine.types;
+
+import java.time.temporal.TemporalUnit;
+import java.util.List;
+
+public class JavaClassWithFilteredMethod {
+    public final int notVisible(List<TemporalUnit> t) {
+        return 0;
+    }
+}